TOREE-390: Lazily initialize spark sessions
This updates the kernel so that it lazily initializes
Spark sessions. This allows the user to configure and
start a session, or to wait until spark is first
referenced to start a session. This shortens the time
to the first interaction with Toree.

It also  adds a notification that Toree
is waiting for a Spark session to start when the
getOrCreate method takes more than 100ms.

Closes #109.

(cherry picked from commit 5cfbc83322c290e84f0cc0848c4eba41508ac54d)

# Conflicts:
#	kernel-api/src/main/scala/org/apache/toree/kernel/api/KernelLike.scala
#	kernel-api/src/main/scala/org/apache/toree/magic/dependencies/IncludeSQLContext.scala
#	kernel/src/main/scala/org/apache/toree/kernel/api/Kernel.scala