[TOREE-437] Establish alternate interrupt handler
Cell interrupts do not take place if Toree is running
in the background (typically due to its parent being
in the background). This PR adds support for the caller
of Toree to specify an alternate interrupt signal via the
environment variable TOREE_ALTERNATE_SIGINT - the value of
which should be the string indicating the alternate signal
name's suffix portion (e.g., "USR2" for SIGUSR2).

The common scenario would be to specify this env value in
the kernel.json's env: stanza when necessary ...

  "display_name": "Apache Toree - Scala",
  "env": {
    "DEFAULT_INTERPRETER": "Scala",
    "PYTHON_EXEC": "python",
    "__TOREE_SPARK_OPTS__": "",
    "__TOREE_OPTS__": "",
    "TOREE_ALTERNATE_SIGINT": "USR2",

If the value of TOREE_ALTERNATE_SIGINT is not recognized
as a valid signal name, a warning message will be logged,
but the kernel's launch will proceed:

17/09/11 08:44:37 [WARN] o.a.t.Main$$anon$1 - Error occurred
establishing alternate signal handler.  Value of
TOREE_ALTERNATE_SIGINT is probably bad: FOO.
Error: Unknown signal: FOO

Closes #140

(cherry picked from commit c2c095962084d8725949f9a395dd0609f0c2a567)