[TOREE-447] Remove mesos-0.18.1-shaded-protobuf.jar
This jar seems to be an old mesos jar that got checked in
into the source repository and does not seem to be in use.
If actually required, it should be added as transient
dependency and listed in license file.

(cherry picked from commit 6d83fe854cbff7fb20a8a6af22fca230fd309c8c)